+++ guizhanlib-minecraft/src/main/java/net/guizhanss/guizhanlib/minecraft/utils/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains utility classes for Minecraft.
+ */
+package net.guizhanss.guizhanlib.minecraft.utils;