+++ CHANGELOG.md
@@ -15,7 +15,9 @@
 
 - 移除简体中文名称相关类（在 GuizhanLibPlugin 中添加）
 - `ItemUtil` 添加 `isValid`、`removeDamage`、`canStack`
-
+- 添加 `MinecraftVersionUtil` 用于判断 Minecraft 版本
+- 添加 `compatibility`
+ 
 ## 1.7.6
 
 - 添加`FileUtil`